+++ openchrom/plugins/net.openchrom.wsd.converter.supplier.gaml/src/net/openchrom/wsd/converter/supplier/gaml/io/ChromatogramReaderVersion120.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Lablicate GmbH.
+ * Copyright (c) 2021, 2023 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -149,7 +149,7 @@ public class ChromatogramReaderVersion120 extends AbstractChromatogramReader imp
 					for(int a = 0; a < absorbances.length; a++) {
 						IVendorScanSignalWSD scanSignal = new VendorScanSignalWSD();
 						scanSignal.setAbundance((float)absorbances[a]);
-						scanSignal.setWavelength(waveLengths[a]);
+						scanSignal.setWavelength((float)waveLengths[a]);
 						scan.addScanSignal(scanSignal);
 					}
 					rt++;
