+++ openchrom/plugins/net.openchrom.xxd.process.supplier.templates/src/net/openchrom/xxd/process/supplier/templates/util/TracesUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018, 2021 Lablicate GmbH.
+ * Copyright (c) 2018, 2023 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -89,7 +89,7 @@ public class TracesUtil {
 			}
 		} else if(scan instanceof IScanWSD) {
 			IScanWSD scanWSD = (IScanWSD)scan;
-			Optional<IScanSignalWSD> optional = scanWSD.getScanSignal((double)trace);
+			Optional<IScanSignalWSD> optional = scanWSD.getScanSignal((float)trace);
 			isTraceContained = optional.isPresent();
 		}
 		//
